refactor: 패키지이름 servlet -> web으로 변경
