+++ src/main/java/woowa/camp/jspcafe/web/UserLoginServlet.java
@@ -1,4 +1,4 @@
-package woowa.camp.jspcafe.servlet;
+package woowa.camp.jspcafe.web;
 
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletConfig;
