+++ src/main/java/woowa/camp/jspcafe/web/UserLogoutServlet.java
@@ -1,4 +1,4 @@
-package woowa.camp.jspcafe.servlet;
+package woowa.camp.jspcafe.web;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
